+++ src/ErgosoftFactory.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace DjThossi\ErgosoftSdk;
 
+use DjThossi\ErgosoftSdk\Api\CancelRippingJobApi;
 use DjThossi\ErgosoftSdk\Api\DeleteJobApi;
 use DjThossi\ErgosoftSdk\Api\GetJobByGuidApi;
 use DjThossi\ErgosoftSdk\Api\GetJobsApi;
@@ -44,6 +45,13 @@ readonly class ErgosoftFactory
     public function createDeleteJobApi(): DeleteJobApi
     {
         return new DeleteJobApi(
+            $this->createClient()
+        );
+    }
+
+    public function createCancelRippingJobApi(): CancelRippingJobApi
+    {
+        return new CancelRippingJobApi(
             $this->createClient()
         );
     }